Refactor start.sh para build projeto when gradle is already installed

+++ start.sh
@@ -2,11 +2,13 @@
 PROJECT_PATH=$(dirname "$0")
 
 installGradle() {
-    cd "$PROJECT_PATH"
-    printf "\n Installing gradle"
-    wget https://services.gradle.org/distributions/gradle-2.4-all.zip
-    unzip gradle-2.4-all.zip
-    rm gradle-2.4-all.zip
+    if [ ! -d "gradle-2.4" ]; then
+        cd "$PROJECT_PATH"
+        printf "\n Installing gradle"
+        wget https://services.gradle.org/distributions/gradle-2.4-all.zip
+        unzip gradle-2.4-all.zip
+        rm gradle-2.4-all.zip
+    fi
 }
 
 run() {
